Move schedule-viewer to a tool, not a test
